Remove unused or unnecessary stuff

+++ commands/rm.js
@@ -2,46 +2,16 @@ var deleteEverythingConfirm = false;
 
 exports.run = (client, message, args) => {
  doNotDelete = false;
-
-if (args.toString() == "bot") {
-const filter = msg => msg.author.bot == true;
-message.delete();
-		message.channel.fetchMessages({
-			limit: 99
-		}).then(messages => {
-			const filteredMessages = messages.filter(filter);
-			message.channel.bulkDelete(filteredMessages);
-			message.channel.send(":white_check_mark: **OK:** I've deleted all messages created by zBot.").then(message => {
-        message.delete({ timeout: 8000 });
-      });
-		})
-	return;
-}
-
-if (args.toString() == "everything") {
-if (deleteEverythingConfirm == false) {
-message.channel.send(":no_entry: **HOLD UP:** Doing this command will result in this channel being **deleted**, then **re-created**. This will remove all messages. If you are aware of what this command will do, you can type `mod:rm everything` again.");
-deleteEverythingConfirm = true;
-message.delete();
-return;
-}
-	message.channel.clone();
-	message.channel.delete();
-
-	var delchannel = message.channel;
-
-	client.channels.get("196793479899250688").sendMessage(":white_check_mark: **OK:** I successfully deleted the channel you were just in and re-created it. All messages in the channel have been removed.").then(message => {
-        message.delete({ timeout: 8000 });
-      });
-	return;
-}
-
 var num = parseInt(args);
 
+if (num > 99) {
+  return message.reply(":no_entry_sign: **NOPE:** Unfortunately, I can't delete more than 99 messages at a time.");
+}
+
 if (num != args) {
-	message.channel.send(":no_entry_sign: **ERROR:** That's not a number...");
+	message.reply(":no_entry_sign: **ERROR:** You need to provide the amount of messages you want to delete.");
 } else {
-	num = num + 1; //Also remove the mod:rm command
+	num = num + 1; //Also remove the rm command
 	message.channel.bulkDelete(num).then(function() {
 		message.channel.send(":white_check_mark: **OK:** I successfully deleted " + eval(num - 1) + " messages.").then(message => {
         message.delete({ timeout: 8000 });
